FE: lock icons added, notifications added; BE: planner fixed

+++ server/services/roomService.js
@@ -1,11 +1,14 @@
 const RoomModel = require('../models/room');
 const MessageModel = require('../models/message');
+const constants = require('./constants');
 
 exports.roomDeleteDb = async (roomId) => {
   try {
     const room = await RoomModel.findById(roomId);
-    room.messages.forEach(async (mId) => await MessageModel.findByIdAndRemove(mId));
-    await RoomModel.findByIdAndRemove(roomId);
+    if (room) {
+      room.messages.forEach(async (mId) => await MessageModel.findByIdAndRemove(mId));
+      await RoomModel.findByIdAndRemove(roomId);
+    }
   } catch(err) {
     console.log(err);
   }
@@ -19,4 +22,78 @@ exports.getRoomsDb = async () => {
     console.log(err);
     return [];
   }
+}
+
+exports.establishRoomKeys = (clients) => {
+  try {
+    let q = "";
+
+    clients.forEach(async (client, i) => {
+      if(i == 0) {
+        q = await qEmitFirst(client);
+      } else if(i == clients.length) {
+        await qEmitLast(client, q);
+
+        let eq = "";
+        [...clients].reverse().forEach((eqClient, eqI) => {
+          if(eqI == 0) {
+            eq = await eqEmitFirst(eqClient);
+          } else if(eqI == clients.length) {
+            await eqEmitLast(eqClient)
+          } else {
+            eq = await eqEmitInter(eqClient, eq)
+          }
+        })
+      } else {
+        q = await qEmitInter(client, q);
+      }
+    })
+
+    return {success: true}
+  } catch(err) {
+    console.log(err);
+    return {success: false}
+  }
+}
+
+const qEmitFirst = (client) => {
+  return new Promise((res, rej) => {
+    client.on('establishResponse', (q) => { res(q) })
+    client.emit('establish', { type: constants.qFirst })
+  })
+}
+
+const qEmitInter = (client, prevQ) => {
+  return new Promise((res, rej) => {
+    client.on('establishResponse', (q) => { res(q) })
+    client.emit('establish', { type: constants.qInter, prevQ })
+  })
+}
+
+const qEmitLast = (client, prevQ) => {
+  return new Promise((res, rej) => {
+    client.on('establishResponse', () => { res() })
+    client.emit('establish', { type: constants.qLast, prevQ })
+  })
+}
+
+const eqEmitFirst = (client) => {
+  return new Promise((res, rej) => {
+    client.on('establishResponse', (eq) => { res(eq) })
+    client.emit('establish', { type: constants.eqFirst })
+  })
+}
+
+const eqEmitInter = (client, prevEq) => {
+  return new Promise((res, rej) => {
+    client.on('establishResponse', (eq) => { res(eq) })
+    client.emit('establish', { type: constants.eqInter, prevEq })
+  })
+}
+
+const eqEmitLast = (client, prevEq) => {
+  return new Promise((res, rej) => {
+    client.on('establishResponse', () => { res() })
+    client.emit('establish', { type: constants.eqLast, prevEq })
+  })
 }
